requirements.txt
Flask==3.1.1
gradio==4.44.0
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.7.0
requests==2.32.4
huggingface-hub==0.33.4

click==8.1.8
filelock==3.18.0
Jinja2==3.1.6
MarkupSafe==2.1.5
Werkzeug==3.1.3
tqdm==4.67.1
packaging==25.0
charset-normalizer==3.4.2
urllib3==2.5.0
idna==3.10
certifi==2025.7.14
PyYAML==6.0.2
regex==2024.11.6
typing_extensions==4.14.1
blinker==1.9.0
itsdangerous==2.2.0
importlib-metadata>=3.6.0

torch==2.2.2
transformers==4.41.2
tokenizers==0.19.1
safetensors==0.5.3

scikit-learn==1.4.2
scipy==1.11.4
joblib==1.5.1
threadpoolctl==3.6.0
networkx==3.2.1

# <-- downgrade to satisfy gradio
pillow==10.4.0

mpmath==1.3.0
sympy==1.14.0
fsspec==2025.5.1
hf-xet==1.1.5
biopython==1.85